refactor: fit in the new RW plugins API
