+++ packages/defer-rw-setup/src/cron/index.ts
@@ -1,0 +1,8 @@
+import { command, builder, description, handler } from "./command.js";
+
+export default {
+  command,
+  handler,
+  description,
+  builder,
+};
